airport-chronicles-incorrect-transcript2v
This video element with no audio, has an incorrect text transcript available on the same page. The transcript contains texts, namely ‘helicopter’ and ‘black’ which are not apparent as you view the video. In the video, there are just a ‘plane’ and ‘black’ cab. But the transcript substituted ‘plane’ with ‘helicopter’ and ‘yellow’ with ‘black.'